+++ .gitignore
@@ -104,5 +104,5 @@ venv.bak/
 .mypy_cache/
 .idea/
 
-# tou
+# onnx
 /out
