+++ docs/tou.md
@@ -5,6 +5,6 @@ ONNX utilties to implement features in onnx-chainer
 Install onnx package as an additional requirements to use this package
 
 ## API
-- `tou.export_testcase`
+- `pytorch_pfn_extras.onnx.export_testcase`
     - ONNX Exportor with testcase outputs like [ONNX testcases](https://github.com/onnx/onnx/tree/master/onnx/backend/test/data/node)
     - Ref: [onnx_chainer.export_testcase](https://docs.chainer.org/en/stable/onnx_chainer/reference/generated/onnx_chainer.export_testcase.html)
