--- pytorch_pfn_extras/tou/__init__.py
@@ -1,2 +1,0 @@
-from pytorch_pfn_extras.tou.export_testcase import export_testcase, is_large_tensor, LARGE_TENSOR_DATA_THRESHOLD  # NOQA
-from pytorch_pfn_extras.tou.annotate import annotate  # NOQA
